Refactored `docker-entrypoint.sh` so `make shell` would be many times faster

+++ container/docker-entrypoint.sh
@@ -10,43 +10,55 @@ export PATH=$HOME/.cargo/bin:$PATH
 # shellcheck disable=SC1091
 source "$HOME/.venv/bin/activate"
 
-# Running connectivity checks
-/srv/scripts/general/connectivity-checks.sh
-
 # Making sure that /ssh-agent has the correct permissions, required mostly for MacOS
 sudo chown -R "$(id -u)":"$(id -g)" /ssh-agent
 
-# KeePass unlocker script
-# Can also be used with ctp-unlock-secrets
+# Secrets unlocker script
+# Can also be used with ctp secrets unlock
 source /srv/scripts/general/secrets-unlock.sh
 
-# Copying mounted certificates to the correct location and trusting them if they are present
-if [ "$(ls -A /tmp/ca-certificates)" ]; then
+# Checking if this is the first run
+# If so then skipping a bunch of steps for better performance
+if [ ! -f /tmp/first-run ]; then
 
-  sudo rsync -ar /tmp/ca-certificates/ /usr/local/share/ca-certificates/ --ignore-existing
-  sudo update-ca-certificates > /dev/null 2>/dev/null # To avoid false positive error rehash: warning: skipping ca-certificates.crt,it does not contain exactly one certificate or CRL
+    # Creating first run file
+    touch /tmp/first-run
+
+    # Running connectivity checks
+    /srv/scripts/general/connectivity-checks.sh
+
+    # Copying mounted certificates to the correct location and trusting them if they are present
+    if [ "$(ls -A /tmp/ca-certificates)" ]; then
+
+      sudo rsync -ar /tmp/ca-certificates/ /usr/local/share/ca-certificates/ --ignore-existing
+      sudo update-ca-certificates > /dev/null 2>/dev/null # To avoid false positive error rehash: warning: skipping ca-certificates.crt,it does not contain exactly one certificate or CRL
+
+    fi
+
+    DOCKER_CONTAINER_ENTRYPOINT_CUSTOM_FILES="/srv/custom/docker-entrypoints/*.sh"
+    for custom_entrypoint in $DOCKER_CONTAINER_ENTRYPOINT_CUSTOM_FILES; do
+      if [ -f "$custom_entrypoint" ]; then
+        # Comment in the echo line below for better debugging
+        # echo -e "\n Processing $custom_entrypoint...\n"
+        # shellcheck disable=SC1090
+        source "$custom_entrypoint"
+      fi
+    done
+
+    DOCKER_CONTAINER_ENTRYPOINT_FILES="/srv/scripts/entrypoints/*.sh"
+    for entrypoint in $DOCKER_CONTAINER_ENTRYPOINT_FILES; do
+      if [ -f "$entrypoint" ]; then
+        # Comment in the echo line below for better debugging
+        # echo -e "\n Processing $entrypoint...\n"
+        # shellcheck disable=SC1090
+        source "$entrypoint"
+      fi
+    done
 
 fi
 
-DOCKER_CONTAINER_ENTRYPOINT_CUSTOM_FILES="/srv/custom/docker-entrypoints/*.sh"
-for custom_entrypoint in $DOCKER_CONTAINER_ENTRYPOINT_CUSTOM_FILES; do
-  if [ -f "$custom_entrypoint" ]; then
-    # Comment in the echo line below for better debugging
-    # echo -e "\n Processing $custom_entrypoint...\n"
-    # shellcheck disable=SC1090
-    source "$custom_entrypoint"
-  fi
-done
-
-DOCKER_CONTAINER_ENTRYPOINT_FILES="/srv/scripts/entrypoints/*.sh"
-for entrypoint in $DOCKER_CONTAINER_ENTRYPOINT_FILES; do
-  if [ -f "$entrypoint" ]; then
-    # Comment in the echo line below for better debugging
-    # echo -e "\n Processing $entrypoint...\n"
-    # shellcheck disable=SC1090
-    source "$entrypoint"
-  fi
-done
+# Running inventory selection script
+source /srv/scripts/general/select-inventory.sh
 
 echo -n -e "${C_RST}"
